requirements.txt
autopep8==2.0.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
idna==3.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.2.0
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
repoze.lru==0.7
requests==2.28.2
Routes==2.5.1
six==1.16.0
tomli==2.0.1
urllib3==1.26.15
